Main

Description

Módulo para a realização do Gloss do projeto de LI1 em 2022/23.

Documentation

data Objeto

Constructors

T Terreno

O Obstaculo

J Jogador

type EstadoGloss = (Jogo, Jogo, Jogo, Bool, Texturas, [Int], Time, ContaJogo, ContaJogo, ContaJogo, Best)

data Texturas

Constructors

Texturas

Fields

player :: [(Jogador, Picture)]

obstaculos :: [(Obstaculo, Picture)]

terrenos :: [(Terreno, Picture)]

menu :: [(Windows, [Picture])]

data Opcoes

Constructors

Jogar

Sair

Oneplayer

Multiplayer

Oneplayer2

Multiplayer2

Progress

Game

GameMultiplayer

GameLost

GameMultiplayerLost

data Windows

Constructors

Menu Opcoes

type AllGame = (Windows, EstadoGloss)

type Time = Int

type ContaJogo = Int

type Best = Int

selecionaLvl :: Int -> [Int] -> Jogo

selecionaLv1 Cria a base do tabuleiro de jogo.

lvl1 :: Mapa -> Int -> [Int] -> Mapa

lvl1 Cria o mapa de jogo.

dm :: Display

dm Dimenciona a janela do jogo.

fr :: Int

fr Indica a Frame rate

reageTempoGloss :: Float -> AllGame -> AllGame

reageTempoGloss Função que reage ao tempo, atualizando o mapa.

reageEventoGloss :: Event -> AllGame -> AllGame

reageEventoGloss Reage aos eventos do Gloss.

estadoGlossInicial :: Texturas -> [Int] -> AllGame

estadoGlossInicial Desenha o estado inicial do Gloss.

desenhaEstadoGloss :: AllGame -> Picture

reageEventoGloss Desenha os diversos eventos do Gloss.

coordenadasLinha :: (Terreno, [Obstaculo]) -> Int -> Int -> [(Objeto, Coordenadas)]

coordenadasLinha Cria a linha do tabuleiro em que associa os objetos da mesma com as suas respetivas coordenadas.

coordenadasMapa :: Mapa -> Int -> [(Objeto, Coordenadas)]

coordenadasMapa Cria uma lista com todas as os objetos associados às respetivas coordenadas.

printGame :: Jogo -> Texturas -> Picture

printGame Faz o print do jogo.

auxPrintGame :: [(Objeto, Coordenadas)] -> Texturas -> [Picture]

auxPrintGame Auxilia da função printGame que tranforma o tabuleiro numa lista de Pictures.

printObjeto :: (Objeto, Coordenadas) -> Texturas -> Picture

printObjeto Associa cada objeto (obstaculo/terreno) à sua Picture.

decodeGame :: Jogo -> [(Objeto, Coordenadas)]

decodeGame Cria uma lista com os objetos e as respetivas coordenadas de todas as peças do tabuleiro.

main :: IO ()

main Main do jogo.