Tarefa1_2022li1g025

Description

Módulo para a realização da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

mapaValido :: Mapa -> Bool

mapaValido Verifica se todas as condiçãoes necessárias para um mapa ser válido são verdadeiras. Para isso,recorre às funções "tarefa1a", "tarefa1b" , "tarefa1c" , "tarefa1d", "tarefa1e" e "tarefa1f".

tarefa1a :: Mapa -> Bool

tarefa1a Verifica se os obstáculos estão nos terrenos que os correspondem.

tarefa1b :: Mapa -> Bool

tarefa1b Verifica se rios contíguos têm direções opostas.

aux1c1 :: [[Obstaculo]] -> Bool

aux1c1 Verifica se não existem mais do que 6 Troncos somados entre o inicio e o fim da linha. Por exemplo: [Tronco,Tronco,Tronco,Nenhum,Tronco,Tronco,Tronco] é uma linha inválida, pois com o movimento do rio eles acabam por ficar os 6 seguidos.

aux1c2 :: [[Obstaculo]] -> Bool

aux1c2 Verifica se não existem mais do que 6 Troncos seguidos.

tarefa1c :: Mapa -> Bool

tarefa1c Chama as funções "aux1c1" e a "aux1c2", sendo True se ambas também o forem.

aux1d :: [Obstaculo] -> Bool

aux1d Verifica se não existem mais do que 4 Carros seguidos.

tarefa1d :: Mapa -> Bool

tarefa1d Chama a função "aux1d" caso a lista de obstáculos seja maior do que 3.

tarefa1e :: Mapa -> Bool

tarefa1e Verifica se existe pelo menos um obstáculo Nenhum na lista de obstáculos.

tarefa1f :: Mapa -> Bool

tarefa1f Verifica se o comprimento da lista de obstáculos de cada linha corresponde exatamente à largura do mapa.

aux1g :: Int -> Terreno -> [(Terreno, [Obstaculo])] -> Int

aux1g A função "aux1g" conta o número de vezes que aparecem seguidos os Rios, ou Estradas ou Relvas.

tarefa1g :: Mapa -> Bool

tarefa1g A função "tarefa1g" retorna False se houverem mais do que 5 Rios seguidos ou mais do que 6 Estradas ou Relvas seguidas.