Tarefa1_2022li1g025_Spec

Description

Módulo para a realização dos testes da Tarefa 1 do projeto de LI1 em 2022/23.

Documentation

testsT1 :: Test

Teste1 Neste caso o mapa fornecido é válido.

Teste2 Mapa incorreto pois temos uma Arvore no local de uma Estrada.

Teste3 Neste caso rios contíguos têm sentidos opostos.

Teste4 Neste caso, os 2 primeiros rios tem sentidos iguais.

Teste5 Neste caso temos mais de 5 troncos seguidos.

Teste6 Neste caso temos mais de 3 carros seguidos.

Teste7 Neste caso temos uma linha do mapa que não têm nenhum obstáculo Nenhum

Teste8 Neste caso, todas as linhas do mapa que têm um obstáculo Nenhum

Teste9 Neste caso temos uma linha do mapa que não tem um número de obstáculos por linha igual igual à largura do mapa.

Teste10 Neste caso temos mais de 4 rios contíguos.

Teste11 Neste caso temos mais de 5 estradas contíguas.