Tarefa2_2022li1g025

Description

Módulo para a realização da Tarefa 2 do projeto de LI1 em 2022/23.

Documentation

terreno1 :: Int -> Int -> Int -> Mapa -> Terreno

terreno1 Auxiliar da função "numero" que vê se o número random se encontra dentro do tamanho da lista dos terrrenos.

numero :: Mapa -> Int -> Int -> Terreno

numero2 Escolhe um terreno para estender o mapa.

obstaculo1 :: Int -> Mapa -> Terreno -> Obstaculo

obstaculo1 Auxiliar da função "mergenumero2" que verifica se o número se encontra dentro do tamanho da lista obstáculos.

numero2 :: Mapa -> Int -> Terreno -> Obstaculo

numero2 Auxiliar da função "mergenumero2" escolhe o número random.

mergenumero2 :: Mapa -> Int -> Int -> Terreno -> [Obstaculo]

mergenumero2 Escolhe obstaculos para formar uma nova linha, dado um terreno.

estendeMapa :: Mapa -> Int -> Int -> Mapa

estendeMapa Atualiza o mapa, colocando uma nova linha em cima.

proximosTerrenosValidos :: Mapa -> Int -> Int -> [Terreno]

proximosTerrenosValidos Vê quais são os próximos terrenos possíveis, tendo em consideração que não podemos ter mais de 4 rios nem 5 estradas consecutivas.

aux2b :: Terreno -> [Obstaculo] -> Int -> Int -> [Obstaculo]

aux2b Auxiliar da função "proximosObstaculos" que vê se não é excedido o tamanho dos carros.

proximosObstaculosValidos :: Int -> (Terreno, [Obstaculo]) -> [Obstaculo]

proximosObstaculosValidos Vê quais são os próximos obstáculos possíveis tendo em conta o terreno que vamos acrescentar ao mapa. Tendo em consideração que não podemos ter mais de 5 troncos nem 3 carros consecutivas, utiliza-se a função auxiliar "aux2b" para verificar essas condições.