Tarefa3_2022li1g025

Description

Módulo para a realização da Tarefa 3 do projeto de LI1 em 2022/23.

Documentation

aux3a :: Jogador -> Int -> [Obstaculo] -> [Obstaculo]

aux3a Função auxiliar da "tarefa3a". Faz o movimento dos troncos e Carros através da combinação dos drops e takes. No caso de um Carro atropelar o jogador, retorna a o mapa no momento do atropelamento.

aux3a1 :: Mapa -> Jogador -> [(Terreno, [Obstaculo])]

aux3a1 Função auxiliar da "tarefa3a". Chama a função "aux3a" para os Rios e Estradas.

tarefa3a :: Jogador -> Mapa -> Mapa

tarefa3a A função "tarefa3a" põem os obstáculos do rio (tronco) e da estrada (carro) a andar com a velocidade e direção do terreno onde estão. Utiliza para isso a "aux3a1" e a "aux3a".

aux3csegunda :: [Obstaculo] -> Jogador -> Int -> Int -> Jogador

aux3csegunda Função auxiliar da "tarefa3c". Faz a movimentação do Jogador.

aux3c :: (Terreno, [Obstaculo]) -> Jogador -> Jogador

aux3c Função auxiliar da "tarefa3c". Chama a função "aux3csegunda" para os Rios.

tarefa3c :: Mapa -> Int -> Jogador -> Jogador

tarefa3c A função "tarefa3c" poem o jogador a se movimentar com o tronco (no caso de estar em cima dele e estar parado). Utiliza as funções "aux3c" e "aux3csegunda".

tarefa3d :: Jogo -> Jogada -> Jogo

tarefa3d A função "tarefa3d" não deixa o jogador ultrapassar os limites do mapa.

aux3 :: Jogo -> Jogada -> Int -> Jogo

aux3 Função auxiliar da "animajogo". Faz a orientação da chamada das várias funções auxiliares.

arvores :: [Obstaculo] -> Int -> Bool

-arvores-- Vẽ se na posição do jogador temos um tronco

auxverArvores :: (Int, Int) -> [(Terreno, [Obstaculo])] -> Bool

-auxVerArvores-- Vẽ se posicao y do jogador e chama a função tronco para ver a posição x.

verLimites :: (Int, Int) -> Mapa -> Bool

-verLimites--

verArvores :: Jogo -> Jogada -> Jogada

-verArvores-- Vê se a jogada coloca o jogador em cima de uma arvore, impedindo que isso aconteça.

animaJogo :: Jogo -> Jogada -> Jogo

animajogo A função "animajogo" permite movimentar os obstáculos (de acordo com a velocidade) do terreno em que se encontram, e o personagem, de acordo com a jogada dada. No caso de o jogador estar em cima do tronco, as suas coordenadas são atualizadas primeiro do que as coordenadas do tronco (pois as coordenadas do tronco são precisas na função "tarefa3a").

animaJogo2 :: Jogo -> Jogada -> Jogo

animajogo2 A função "animajogo" permite movimentar os obstáculos (de acordo com a velocidade) do terreno em que se encontram.