Tarefa4_2022li1g025_Spec

Description

Módulo para a realização dos testes da Tarefa 4 do projeto de LI1 em 2022/23.

Documentation

testsT4 :: Test

Teste 1 Verifica que, se o jogador for para fora do mapa, o jogo termina.

Teste 2 Verifica que, se o jogador cair na água, o jogo termina.

Teste 3 Verifica que, se o jogador for atingido por um carro, o jogo termina.

Teste 4 Verifica que, se o jogador estiver numa posição válida , o jogo continua.